Move g1- & g2-serialization specific methods to mod util
